+++ src/app.js
@@ -1,6 +1,23 @@
 /* OK BUTTON
 OK button is disabled by default */
 
+async function retrieveNames(numberOfStudents) {
+    return fetch('https://random-data-api.com/api/name/random_name?size=' + numberOfStudents)
+        .then(response => response.json())
+        // Functional way of iterate the array
+        .then(data =>
+            data.map(element => element.name)
+        )
+        // Another way of iterating the array (non functional)
+        // .then(data => {
+        //     let students = [];
+        //     for (const element of data) {
+        //         students.push(element.name);
+        //     }
+        //   return students;
+        // })
+}
+
 let buttonOk = document.getElementById("studentsButton");
 
 buttonOk.disabled = true;
@@ -104,14 +121,19 @@ function clearAndCompleteTable() {
         table.deleteRow(i);
     }
     students = []; // clear array
-    for (let i = 1; i <= numberOfStudentsInput.value; i++) {
-        addRowToStudentsTable(i);
-    }
+    retrieveNames(numberOfStudentsInput.value)
+        .then(names => {
+            for (let i = 0; i < names.length; i++) {
+                let name = names[i];
+                addRowToStudentsTable(i + 1, name);
+            }
+        })
+
 };
 
 /* Add rows to students table depending on the input number */
 
-function addRowToStudentsTable(i) {
+function addRowToStudentsTable(i, name) {
     let row = table.insertRow(i);
     let cell1 = row.insertCell(0);
     let cell2 = row.insertCell(1);
@@ -120,9 +142,9 @@ function addRowToStudentsTable(i) {
     // row.insertCell(3);
 
     cell1.innerHTML = i; // Number
-    cell2.innerHTML = "Elvis Presley";
-    let student = new Student(i, "Elvis Presley");
-    students.push(student);
+    cell2.innerHTML = name;
+    let student = new Student(i, name);
+    students.push(student); //Add student to the array
 };
 
 /* Send button*/
@@ -152,6 +174,7 @@ function confirmationMessage() {
     location.reload(); // Page reloads after clicking on OK
 };
 
+/* Array of students ready to be sent */
 class Student {
     constructor(id, name) {
         this.id = id;